demo-app/requirements.txt
flask==3.1.1
flask-cors==5.0.1
anthropic==0.49.0
python-dotenv==1.1.0
